Arreglada lista de regalos

+++ SmartTrade/SmartTrade/ViewModels/GiftsModel.cs
@@ -21,45 +21,86 @@ public class GiftsModel : ViewModelBase
     public bool? RemoveButtonIsVisible { get; set; }
     public int ComboBoxIndex {  get; set; }
     public ObservableCollection<GiftItemModel> Gifts { get; set; }
-    List<GiftListDTO>? GiftLists => Service.GiftLists;
+    public List<GiftListDTO>? GiftLists => Service.GiftLists;
     public ObservableCollection<string> GiftListsNames { get; set; }
 
     public GiftsModel()
     {
         Gifts = new ObservableCollection<GiftItemModel>();
         GiftListsNames = new ObservableCollection<string>();
-        ComboBoxIndex = 0;
 
-        Service.OnGiftListsChanged += UpdateView;
-        Service.OnGiftsChanged += UpdateView;
         Service.OnGiftsChanged += Calculate;
-        Calculate();
-        UpdateView();
     }
 
-    ~GiftsModel()
+    public void UpdateView(int selectedIndex)
     {
-        
-    }
-
-    public void UpdateView()
-    {
-        GiftListsNames = new ObservableCollection<string>();
+        Gifts.Clear();
 
         if (Service.GiftLists == null) { return; }
 
-        foreach (GiftListDTO giftList in this.GiftLists)
+        if (GiftLists.Count > selectedIndex && selectedIndex > -1)
         {
-            GiftListsNames.Add(giftList.Name);
-        }
+            ComboBoxIndex = selectedIndex;
 
-        if (GiftListsNames.Count == 0) { EditButtonIsVisible = false; RemoveButtonIsVisible = false; }
-        else { 
-        foreach (GiftDTO gift in GiftLists[ComboBoxIndex].Gifts)
+            foreach (GiftDTO gift in GiftLists[selectedIndex].Gifts)
             {
                 Gifts.Add(new GiftItemModel(gift, this));
             }
         }
+        else ComboBoxIndex = 0;
+
+        if (GiftListsNames.Count == 0) { EditButtonIsVisible = false; RemoveButtonIsVisible = false; }
+        else {
+            EditButtonIsVisible = true; RemoveButtonIsVisible = true;
+        }
+
+        if (Gifts.Count == 0) { NoElementsIsVisible = true; NoElementsText = "No elements in the gift list"; }
+        else { NoElementsIsVisible = false;}
+
+        RaisePropertyChanges();
+        Calculate();
+    }
+
+    public void UpdateView()
+    {
+        GiftListsNames.Clear();
+        Gifts.Clear();
+
+        if (Service.GiftLists == null) { return; }
+
+        foreach (GiftListDTO giftList in GiftLists)
+        {
+            GiftListsNames.Add(giftList.Name);
+        }
+
+        if (GiftLists.Count > ComboBoxIndex && ComboBoxIndex > -1)
+        {
+            foreach (GiftDTO gift in GiftLists[ComboBoxIndex].Gifts)
+            {
+                Gifts.Add(new GiftItemModel(gift, this));
+            }
+        }
+
+        if (GiftListsNames.Count == 0) { EditButtonIsVisible = false; RemoveButtonIsVisible = false; }
+        else
+        {
+            EditButtonIsVisible = true; RemoveButtonIsVisible = true;
+        }
+
+        if (Gifts.Count == 0) { NoElementsIsVisible = true; NoElementsText = "No elements in the gift list"; }
+        else { NoElementsIsVisible = false; }
+
+        RaisePropertyChanges();
+        Calculate();
+    }
+
+    private void RaisePropertyChanges()
+    {
+        this.RaisePropertyChanged(nameof(ComboBoxIndex));
+        this.RaisePropertyChanged(nameof(EditButtonIsVisible));
+        this.RaisePropertyChanged(nameof(RemoveButtonIsVisible));
+        this.RaisePropertyChanged(nameof(NoElementsIsVisible));
+        this.RaisePropertyChanged(nameof(NoElementsText));
     }
 
     public void Calculate()
@@ -80,23 +121,41 @@ public class GiftsModel : ViewModelBase
         this.RaisePropertyChanged(nameof(SubTotal));
         this.RaisePropertyChanged(nameof(ShippingCost));
         this.RaisePropertyChanged(nameof(Total));
-    }
 
-    public void UnSubscribeFromGiftsNotifications()
-    {
-        Service.OnGiftListsChanged -= UpdateView;
-        Service.OnGiftsChanged -= UpdateView;
-        Service.OnGiftsChanged -= Calculate;
+        if (Gifts.Count == 0)
+        {
+            if (Gifts.Count == 0) { NoElementsIsVisible = true; NoElementsText = "No elements in the gift list"; }
+            else { NoElementsIsVisible = false; }
+        }
+
+        this.RaisePropertyChanged(nameof(NoElementsIsVisible));
+        this.RaisePropertyChanged(nameof(NoElementsText));
     }
 
     public void AddGiftList(string name, DateOnly? date)
     {
         if(GiftLists.FindIndex(gl => gl.Name == name) != -1) { throw new Exception("A gift list with the same name already exists"); }
         Service.AddGiftListAsync(name,date);
+        UpdateView();
     }
 
     public void RemoveGiftList()
     { 
+        if(ComboBoxIndex == -1) { return; }
+
         Service.RemoveGiftListAsync( GiftLists[ComboBoxIndex].Name);
+        UpdateView();
+    }
+
+    public void EditGiftList(string name, string newName, DateOnly? date)
+    {
+        if (GiftLists.FindIndex(gl => gl.Name == newName && gl.Name != name) != -1) { throw new Exception("A gift list with the same name already exists"); }
+        Service.EditGiftListAsync(name, newName, date);
+        UpdateView();
+    }
+
+    public void Dispose()
+    {
+        Service.OnGiftsChanged -= Calculate;
     }
 }
